Correct inheritance
